refactor: rename scraper.py → fetcher.py, remove scraping references
The module uses the GovInfo API, not HTML scraping. Rename the file and
update all imports, the test file, and the README to use "fetcher" and
reflect the GovInfo CPD data source.
